Time: 49 ms (39.04%), Space: 113.7 MB (27.63%) - LeetHub

+++ 2171-removing-minimum-number-of-magic-beans/2171-removing-minimum-number-of-magic-beans.cpp
@@ -11,7 +11,7 @@ public:
             prefixSum[i]=prefixSum[i-1]+beans[i];
         }
 
-        long long ans=prefixSum[n-1]-(n)*beans[0];
+        long long ans= prefixSum[n-1]-(long long)(n)*beans[0];
         long long sum=beans[0];
         for(int i=1;i<n;i++){
             long long temp = sum+(prefixSum[n-1]-prefixSum[i-1])-(long long)(n-i)*beans[i];
